+++ src/sklearndf/regression/_regression_v0_23.py
@@ -6,9 +6,6 @@ import logging
 
 from sklearn.linear_model import GammaRegressor, PoissonRegressor, TweedieRegressor
 
-# noinspection PyProtectedMember
-from sklearn.linear_model._glm import GeneralizedLinearRegressor
-
 from pytools.api import AllTracker
 
 from ..wrapper import RegressorWrapperDF
@@ -17,7 +14,6 @@ log = logging.getLogger(__name__)
 
 __all__ = [
     "GammaRegressorDF",
-    "GeneralizedLinearRegressorDF",
     "PoissonRegressorDF",
     "TweedieRegressorDF",
 ]
@@ -51,12 +47,6 @@ class GammaRegressorDF(RegressorWrapperDF[GammaRegressor], native=GammaRegressor
 
 class TweedieRegressorDF(RegressorWrapperDF[TweedieRegressor], native=TweedieRegressor):
     """Stub for DF wrapper of class ``TweedieRegressor``"""
-
-
-class GeneralizedLinearRegressorDF(
-    RegressorWrapperDF[GeneralizedLinearRegressor], native=GeneralizedLinearRegressor
-):
-    """Stub for DF wrapper of class ``GeneralizedLinearRegressor``"""
 
 
 #