+++ RELEASE_NOTES.rst
@@ -7,10 +7,48 @@ Release Notes
 .. |nbsp| unicode:: 0xA0
    :trim:
 
+*sklearndf* 2.2
+---------------
+
+*sklearndf* 2.2 adds support for
+`scikit-learn |nbsp| 1.2 <https://scikit-learn.org/1.2>`_, and enhances the EstimatorDF
+API.
+
+2.2.0
+~~~~~
+
+- API: new property :attr:`.EstimatorDF.output_names_` to get the names of the output
+  columns the estimator was fitted with.
+
+
+*sklearndf* 2.1
+---------------
+
+*sklearndf* 2.1 adds support for
+`scikit-learn |nbsp| 1.1 <https://scikit-learn.org/1.1>`_.
+
+2.1.0
+~~~~~
+
+- API: new clusterer :class:`.BisectingKMeansDF`
+- API: new transformer :class:`.MiniBatchNMFDF`
+- API: new transformer :class:`.RandomTreesEmbeddingDF`; note that class
+  :class:`~sklearn.ensemble.RandomTreesEmbedding` existed previously in *scikit-learn*,
+  but is based on :class:`~sklearn.base.TransformerMixin` only as of
+  *scikit-learn* |nbsp| 1.1
+- API: support parameters ``max_categories`` and ``min_frequency`` of
+  :class:`.OneHotEncoderDF`, introduced in *scikit-learn* |nbsp| 1.1
+- API: support array-like values for parameter ``drop`` of :class:`.OneHotEncoderDF`
+- API: support ``"passthrough"`` as a transformer in :class:`.FeatureUnionDF`
+- API: remove ``GeneralizedLinearRegressorDF`` since the underlying native estimator is
+  a base class and not intended to be used as a regressor of its own
+
+
 *sklearndf* 2.0
 ---------------
 
-*sklearndf* 2 adds support for `scikit-learn 1.0 <https://scikit-learn.org/1.0>`_,
+*sklearndf* 2.0 adds support for
+`scikit-learn |nbsp| 1.0 <https://scikit-learn.org/1.0>`_,
 adds data frame support for clusterers along with additional API enhancements and
 improvements, and is now subject to static type checking with |mypy|.